[examples] rename files minimalradium -> AllPrimitivesComponent
