+++ examples/DrawPrimitives/AllPrimitivesComponent.hpp
